requirements.txt
-e git+https://github.com/ckan/ckan.git@ckan-2.8.6#egg=ckan
-e git+https://github.com/GSA/ckanext-googleanalyticsbasic#egg=ckanext-googleanalyticsbasic
-e git+https://github.com/GSA/USMetadata.git@ckan-2-8#egg=ckanext-usmetadata
-e git+https://github.com/GSA/ckanext-datajson.git@master#egg=ckanext-datajson
-e git+https://github.com/GSA/ckanext-saml2.git@datagov/v0.4.0#egg=ckanext-saml2
-e git+https://github.com/GSA/webob.git@ckan-patch#egg=webob
ckanext-dcat-usmetadata~=0.2
-e git+https://github.com/GSA/ckanext-s3filestore@datagov-v0.1.1#egg=ckanext-s3filestore
-e ./ckanext-datagov_inventory
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file requirements=py2.txt requirements-py2.in
#

# From root CKAN
argparse==1.4.0           # via ofs
Babel==2.3.4              # via flask-babel
Beaker==1.9.0             # via pylons
bleach==3.1.4
click==6.7
decorator==4.2.1          # via pylons, sqlalchemy-migrate
fanstatic==0.12
Flask-Babel==0.11.2
Flask==0.12.4             # via flask-babel
FormEncode==1.3.1         # via Pylons
funcsigs==1.0.2           # via beaker
html5lib==1.0.1           # via bleach
itsdangerous==0.24        # via flask
Jinja2==2.10.1            # via flask, flask-babel
Mako==1.0.7               # via pylons
Markdown==2.6.7
MarkupSafe==1.0           # via jinja2, mako, webhelpers
mock==1.0.1               # needed for usmetadata
newrelic==5.20.1.150
nose==1.3.7               # via pylons
ofs==0.4.2
Pairtree==0.7.1-T
passlib==1.7.3
paste==1.7.5.1
PasteDeploy==1.5.2        # via pastescript, pylons
PasteScript==2.0.2        # via pylons
pbr==1.10.0               # via sqlalchemy-migrate
polib==1.0.7
psycopg2==2.7.3.2
Pygments==2.2.0           # via weberror
Pylons==0.9.7
pysolr==3.6.0
python-dateutil==1.5
python-magic==0.4.15
pytz==2016.7
pyutilib.component.core==4.6.4
redis==2.10.6             # via rq
repoze.lru==0.7           # via routes
repoze.who-friendlyform==1.0.8
repoze.who==2.3
requests==2.20
Routes==1.13              # via pylons
rq==0.6.0
simplejson==3.10.0
six==1.11.0               # via bleach, html5lib, pastescript, pyutilib.component.core, sqlalchemy-migrate
sqlalchemy-migrate==0.10.0
SQLAlchemy==1.1.11       # via sqlalchemy-migrate
sqlparse==0.2.2
Tempita==0.5.2            # via pylons, sqlalchemy-migrate, weberror
tzlocal==1.3
unicodecsv==0.14.1
urllib3==1.25.9
vdm==0.14
webencodings==0.5.1       # via html5lib
WebError==0.13.1          # via pylons
WebHelpers==1.3           # via pylons
WebTest==1.4.3            # via pylons
Werkzeug==0.15.3         # via flask
zope.interface==4.3.2

# ckanext-saml2 dependencies
defusedxml==0.6.0
M2Crypto
python-memcached==1.48
-e git+https://github.com/GSA/pysaml2.git@datagov/v4.9.0#egg=pysaml2

#datajson
jsonschema~=2.4.0
rfc3987
pyyaml
lepl

# ckanext-s3filestore
boto3>=1.4.4
ckantoolkit

#Serve
gunicorn==19.10.0
gevent
